Adding updated webtoaff haskell script.
